Add files via upload
This was a comprehensive group project where each member contributed and built a report for two investors with different investment expectations. Here we used the Python libraries like Pandas, Numpy, Matplotlib for analysis and visualization. I contributed in the form of risk analysis in the Jupytor notebook and building PowerBI report.